renamed packages from offlinemoney to offlinedigitaleuro
