+++ offlinedigitaleuro/build.gradle
@@ -34,7 +34,7 @@ android {
 //    kotlinOptions {
 //        jvmTarget = '1.8'
 //    }
-    namespace 'nl.tudelft.trustchain.offlinemoney'
+    namespace 'nl.tudelft.trustchain.offlinedigitaleuro'
 }
 
 dependencies {
