+++ offlinedigitaleuro/src/androidTest/java/nl/tudelft/trustchain/offlinedigitaleuro/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package nl.tudelft.trustchain.offlinemoney
+package nl.tudelft.trustchain.offlinedigitaleuro
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
