+++ offlinedigitaleuro/src/main/java/nl/tudelft/trustchain/offlinedigitaleuro/src/Wallet.kt
@@ -1,11 +1,9 @@
-package nl.tudelft.trustchain.offlinemoney.src
+package nl.tudelft.trustchain.offlinedigitaleuro.src
 
-import mu.KotlinLogging
 import nl.tudelft.ipv8.keyvault.JavaCryptoProvider
 import nl.tudelft.ipv8.keyvault.defaultCryptoProvider
 import nl.tudelft.ipv8.keyvault.PrivateKey
 import nl.tudelft.ipv8.keyvault.PublicKey
-import java.security.SecureRandom
 
 class Wallet {
     val privateKey: PrivateKey = defaultCryptoProvider.generateKey()
