+++ offlinedigitaleuro/src/main/res/layout/activity_main_offline_money.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MainActivityOfflineMoney">
+    tools:context=".MainActivityOfflineDigitalEuro">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/balanceLayout"
