+++ offlinedigitaleuro/src/main/res/menu/offlinedigitaleuro_menu.xml
